User model and authentication added.

+++ backend/controllers/goalController.js
@@ -1,11 +1,12 @@
 const asyncHandler = require('express-async-handler')
 const Goal = require('../models/goalModel')
+const User = require('../models/userModel')
 
 // @desc    Get goals
 // @route   GET /api/goals
 // @access  Private
 const getGoals = asyncHandler(async (req, res) => {
-    const goals = await Goal.find()
+    const goals = await Goal.find({user: req.user.id})
     res.status(200).json(goals)
 })
 
@@ -19,7 +20,8 @@ const setGoal = asyncHandler(async (req, res) => {
     }
 
     const goal = await Goal.create({
-        text: req.body.text
+        text: req.body.text,
+        user: req.user.id
     })
     res.status(200).json(goal)
 })
@@ -34,6 +36,18 @@ const updateGoal = asyncHandler(async (req, res) => {
     if (!goal){
         res.status(400)
         throw new Error("Goal not found.")
+    }
+
+    const user = await User.findById(req.user.id)
+
+    if(!user) {
+        res.status(401)
+        throw new Error("User not found")
+    }
+
+    if(goal.user.toString() != user.id){
+        res.status(401)
+        throw new Error("User not authorized")
     }
 
     const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
@@ -53,6 +67,18 @@ const deleteGoal = asyncHandler(async (req, res) => {
     if (!goal){
         res.status(400)
         throw new Error("Goal not found.")
+    }
+
+    const user = await User.findById(req.user.id)
+
+    if(!user) {
+        res.status(401)
+        throw new Error("User not found")
+    }
+
+    if(goal.user.toString() != user.id){
+        res.status(401)
+        throw new Error("User not authorized")
     }
 
     const deletedGoal = await Goal.findByIdAndDelete(req.params.id)
